requirements.txt
django
django-filter
djangorestframework
pydicom
numpy
